Move files to where they belong to
