+++ frontend/src/metabase/admin/databases/containers/DatabaseListApp/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DatabaseListApp";
